# kratos-server/api/rulego/v1/component_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.5.3
// - protoc             v3.21.9
// source: rulego/v1/component.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationComponentServiceGetCustomComponentDSL = "/api.rulego.v1.ComponentService/GetCustomComponentDSL"
const OperationComponentServiceListComponents = "/api.rulego.v1.ComponentService/ListComponents"
const OperationComponentServiceListCustomComponents = "/api.rulego.v1.ComponentService/ListCustomComponents"
const OperationComponentServiceListMarketplaceComponents = "/api.rulego.v1.ComponentService/ListMarketplaceComponents"
const OperationComponentServiceListSharedComponents = "/api.rulego.v1.ComponentService/ListSharedComponents"
const OperationComponentServiceUninstallCustomComponent = "/api.rulego.v1.ComponentService/UninstallCustomComponent"
const OperationComponentServiceUpgradeCustomComponent = "/api.rulego.v1.ComponentService/UpgradeCustomComponent"

type ComponentServiceHTTPServer interface {
	GetCustomComponentDSL(context.Context, *GetCustomComponentDSLRequest) (*GetCustomComponentDSLResponse, error)
	ListComponents(context.Context, *ListComponentsRequest) (*ListComponentsResponse, error)
	ListCustomComponents(context.Context, *ListCustomComponentsRequest) (*ListCustomComponentsResponse, error)
	ListMarketplaceComponents(context.Context, *ListMarketplaceComponentsRequest) (*ListMarketplaceComponentsResponse, error)
	ListSharedComponents(context.Context, *ListSharedComponentsRequest) (*ListSharedComponentsResponse, error)
	UninstallCustomComponent(context.Context, *UninstallCustomComponentRequest) (*UninstallCustomComponentResponse, error)
	UpgradeCustomComponent(context.Context, *UpgradeCustomComponentRequest) (*UpgradeCustomComponentResponse, error)
}

func RegisterComponentServiceHTTPServer(s *http.Server, srv ComponentServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/components", _ComponentService_ListComponents0_HTTP_Handler(srv))
	r.GET("/api/v1/dynamic-components", _ComponentService_ListCustomComponents0_HTTP_Handler(srv))
	r.GET("/api/v1/dynamic-components/{id}", _ComponentService_GetCustomComponentDSL0_HTTP_Handler(srv))
	r.POST("/api/v1/dynamic-components/{id}", _ComponentService_UpgradeCustomComponent0_HTTP_Handler(srv))
	r.DELETE("/api/v1/dynamic-components/{id}", _ComponentService_UninstallCustomComponent0_HTTP_Handler(srv))
	r.GET("/api/v1/shared-nodes", _ComponentService_ListSharedComponents0_HTTP_Handler(srv))
	r.GET("/api/v1/marketplace/components", _ComponentService_ListMarketplaceComponents0_HTTP_Handler(srv))
}

func _ComponentService_ListComponents0_HTTP_Handler(srv ComponentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListComponentsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationComponentServiceListComponents)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListComponents(ctx, req.(*ListComponentsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListComponentsResponse)
		return ctx.Result(200, reply)
	}
}

func _ComponentService_ListCustomComponents0_HTTP_Handler(srv ComponentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListCustomComponentsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationComponentServiceListCustomComponents)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListCustomComponents(ctx, req.(*ListCustomComponentsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListCustomComponentsResponse)
		return ctx.Result(200, reply)
	}
}

func _ComponentService_GetCustomComponentDSL0_HTTP_Handler(srv ComponentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetCustomComponentDSLRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationComponentServiceGetCustomComponentDSL)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetCustomComponentDSL(ctx, req.(*GetCustomComponentDSLRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetCustomComponentDSLResponse)
		return ctx.Result(200, reply)
	}
}

func _ComponentService_UpgradeCustomComponent0_HTTP_Handler(srv ComponentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpgradeCustomComponentRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationComponentServiceUpgradeCustomComponent)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpgradeCustomComponent(ctx, req.(*UpgradeCustomComponentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpgradeCustomComponentResponse)
		return ctx.Result(200, reply)
	}
}

func _ComponentService_UninstallCustomComponent0_HTTP_Handler(srv ComponentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UninstallCustomComponentRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationComponentServiceUninstallCustomComponent)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UninstallCustomComponent(ctx, req.(*UninstallCustomComponentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UninstallCustomComponentResponse)
		return ctx.Result(200, reply)
	}
}

func _ComponentService_ListSharedComponents0_HTTP_Handler(srv ComponentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListSharedComponentsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationComponentServiceListSharedComponents)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListSharedComponents(ctx, req.(*ListSharedComponentsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListSharedComponentsResponse)
		return ctx.Result(200, reply)
	}
}

func _ComponentService_ListMarketplaceComponents0_HTTP_Handler(srv ComponentServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListMarketplaceComponentsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationComponentServiceListMarketplaceComponents)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListMarketplaceComponents(ctx, req.(*ListMarketplaceComponentsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListMarketplaceComponentsResponse)
		return ctx.Result(200, reply)
	}
}

type ComponentServiceHTTPClient interface {
	GetCustomComponentDSL(ctx context.Context, req *GetCustomComponentDSLRequest, opts ...http.CallOption) (rsp *GetCustomComponentDSLResponse, err error)
	ListComponents(ctx context.Context, req *ListComponentsRequest, opts ...http.CallOption) (rsp *ListComponentsResponse, err error)
	ListCustomComponents(ctx context.Context, req *ListCustomComponentsRequest, opts ...http.CallOption) (rsp *ListCustomComponentsResponse, err error)
	ListMarketplaceComponents(ctx context.Context, req *ListMarketplaceComponentsRequest, opts ...http.CallOption) (rsp *ListMarketplaceComponentsResponse, err error)
	ListSharedComponents(ctx context.Context, req *ListSharedComponentsRequest, opts ...http.CallOption) (rsp *ListSharedComponentsResponse, err error)
	UninstallCustomComponent(ctx context.Context, req *UninstallCustomComponentRequest, opts ...http.CallOption) (rsp *UninstallCustomComponentResponse, err error)
	UpgradeCustomComponent(ctx context.Context, req *UpgradeCustomComponentRequest, opts ...http.CallOption) (rsp *UpgradeCustomComponentResponse, err error)
}

type ComponentServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewComponentServiceHTTPClient(client *http.Client) ComponentServiceHTTPClient {
	return &ComponentServiceHTTPClientImpl{client}
}

func (c *ComponentServiceHTTPClientImpl) GetCustomComponentDSL(ctx context.Context, in *GetCustomComponentDSLRequest, opts ...http.CallOption) (*GetCustomComponentDSLResponse, error) {
	var out GetCustomComponentDSLResponse
	pattern := "/api/v1/dynamic-components/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationComponentServiceGetCustomComponentDSL))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *ComponentServiceHTTPClientImpl) ListComponents(ctx context.Context, in *ListComponentsRequest, opts ...http.CallOption) (*ListComponentsResponse, error) {
	var out ListComponentsResponse
	pattern := "/api/v1/components"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationComponentServiceListComponents))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *ComponentServiceHTTPClientImpl) ListCustomComponents(ctx context.Context, in *ListCustomComponentsRequest, opts ...http.CallOption) (*ListCustomComponentsResponse, error) {
	var out ListCustomComponentsResponse
	pattern := "/api/v1/dynamic-components"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationComponentServiceListCustomComponents))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *ComponentServiceHTTPClientImpl) ListMarketplaceComponents(ctx context.Context, in *ListMarketplaceComponentsRequest, opts ...http.CallOption) (*ListMarketplaceComponentsResponse, error) {
	var out ListMarketplaceComponentsResponse
	pattern := "/api/v1/marketplace/components"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationComponentServiceListMarketplaceComponents))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *ComponentServiceHTTPClientImpl) ListSharedComponents(ctx context.Context, in *ListSharedComponentsRequest, opts ...http.CallOption) (*ListSharedComponentsResponse, error) {
	var out ListSharedComponentsResponse
	pattern := "/api/v1/shared-nodes"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationComponentServiceListSharedComponents))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *ComponentServiceHTTPClientImpl) UninstallCustomComponent(ctx context.Context, in *UninstallCustomComponentRequest, opts ...http.CallOption) (*UninstallCustomComponentResponse, error) {
	var out UninstallCustomComponentResponse
	pattern := "/api/v1/dynamic-components/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationComponentServiceUninstallCustomComponent))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *ComponentServiceHTTPClientImpl) UpgradeCustomComponent(ctx context.Context, in *UpgradeCustomComponentRequest, opts ...http.CallOption) (*UpgradeCustomComponentResponse, error) {
	var out UpgradeCustomComponentResponse
	pattern := "/api/v1/dynamic-components/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationComponentServiceUpgradeCustomComponent))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}