# kratos-server/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/components:
        get:
            tags:
                - ComponentService
            description: 获取所有组件列表
            operationId: ComponentService_ListComponents
            parameters:
                - name: username
                  in: query
                  schema:
                    type: string
                - name: category
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rulego.v1.ListComponentsResponse'
    /api/v1/dynamic-components:
        get:
            tags:
                - ComponentService
            description: 获取自定义动态组件列表
            operationId: ComponentService_ListCustomComponents
            parameters:
                - name: username
                  in: query
                  schema:
                    type: string
                - name: checkMarketplace
                  in: query
                  schema:
                    type: boolean
                - name: checkMy
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rulego.v1.ListCustomComponentsResponse'
    /api/v1/dynamic-components/{id}:
        get:
            tags:
                - ComponentService
            description: 获取自定义动态组件DSL
            operationId: ComponentService_GetCustomComponentDSL
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: username
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rulego.v1.GetCustomComponentDSLResponse'
        post:
            tags:
                - ComponentService
            description: 安装/升级自定义动态组件
            operationId: ComponentService_UpgradeCustomComponent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rulego.v1.UpgradeCustomComponentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rulego.v1.UpgradeCustomComponentResponse'
        delete:
            tags:
                - ComponentService
            description: 卸载自定义动态组件
            operationId: ComponentService_UninstallCustomComponent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: username
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rulego.v1.UninstallCustomComponentResponse'
...
    /api/v1/marketplace/components:
        get:
            tags:
                - ComponentService
            description: 获取组件市场组件列表
            operationId: ComponentService_ListMarketplaceComponents
            parameters:
                - name: username
                  in: query
                  schema:
                    type: string
                - name: checkMy
                  in: query
                  schema:
                    type: boolean
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rulego.v1.ListMarketplaceComponentsResponse'
...
    /api/v1/shared-nodes:
        get:
            tags:
                - ComponentService
            description: 获取所有共享组件
            operationId: ComponentService_ListSharedComponents
            parameters:
                - name: username
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rulego.v1.ListSharedComponentsResponse'
...
components:
    schemas:
        api.rulego.v1.ComponentInfo:
            type: object
            properties:
                type:
                    type: string
                name:
                    type: string
                category:
                    type: string
                description:
                    type: string
                version:
                    type: string
                author:
                    type: string
                configurationForm:
                    type: object
                    additionalProperties:
                        type: string
                isCustom:
                    type: boolean
                isInstalled:
                    type: boolean
            description: 组件信息
...
        api.rulego.v1.CustomComponentDetail:
            type: object
            properties:
                id:
                    type: string
                type:
                    type: string
                name:
                    type: string
                description:
                    type: string
                version:
                    type: string
                author:
                    type: string
                dsl:
                    type: string
                createdTime:
                    type: string
                updatedTime:
                    type: string
                isInstalled:
                    type: boolean
            description: 自定义组件详情
...
        api.rulego.v1.GetCustomComponentDSLResponse:
            type: object
            properties:
                component:
                    $ref: '#/components/schemas/api.rulego.v1.CustomComponentDetail'
            description: 获取自定义组件DSL响应
...
        api.rulego.v1.ListComponentsResponse:
            type: object
            properties:
                endpoints:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rulego.v1.ComponentInfo'
                nodes:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rulego.v1.ComponentInfo'
                builtins:
                    type: object
                    additionalProperties:
                        type: string
                sharedNodes:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rulego.v1.SharedComponentInfo'
            description: 获取组件列表响应
        api.rulego.v1.ListCustomComponentsResponse:
            type: object
            properties:
                components:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rulego.v1.CustomComponentDetail'
            description: 获取自定义组件列表响应
        api.rulego.v1.ListMarketplaceComponentsResponse:
            type: object
            properties:
                components:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rulego.v1.CustomComponentDetail'
            description: 获取组件市场组件列表响应
...
        api.rulego.v1.ListSharedComponentsResponse:
            type: object
            properties:
                components:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rulego.v1.SharedComponentInfo'
            description: 获取共享组件列表响应
...
        api.rulego.v1.SharedComponentInfo:
            type: object
            properties:
                id:
                    type: string
                type:
                    type: string
                name:
                    type: string
                description:
                    type: string
                endpointType:
                    type: string
                configuration:
                    type: object
                    additionalProperties:
                        type: string
            description: 共享组件信息
        api.rulego.v1.UninstallCustomComponentResponse:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
            description: 卸载自定义组件响应
...
        api.rulego.v1.UpgradeCustomComponentRequest:
            type: object
            properties:
                id:
                    type: string
                dsl:
                    type: string
                username:
                    type: string
            description: 安装/升级自定义组件请求
        api.rulego.v1.UpgradeCustomComponentResponse:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
            description: 安装/升级自定义组件响应
...